fix: move types const into constant.ts
