+++ packages/hmr/lib/types.ts
@@ -1,7 +1,4 @@
-export const DO_UPDATE = 'do_update';
-export const DONE_UPDATE = 'done_update';
-export const BUILD_COMPLETE = 'build_complete';
-export const ERROR = 'error';
+import { BUILD_COMPLETE, DO_UPDATE, DONE_UPDATE, ERROR } from './constant';
 
 type UpdateRequestMessage = {
   type: typeof DO_UPDATE;
